+++ src/frontend/global_signals.py
@@ -33,6 +33,10 @@ class GlobalSignals(QObject):
     wizard_process_btn_clicked = Signal()
     wizard_process_btn_change_txt = Signal(str)
     wizard_process_btn_set_hidden = Signal()
+
+    # process
+    prompt_tokens_response = Signal(object)  # dict[str, str]
+    overview_prompt_response = Signal(object)  # dict[TrackerSelection, dict[str | None, str]]
     ########### SIGNALS ###########
 
     def __new__(cls, *args, **kwargs):